feat(data_handling): change process output to csv
Change output of process to csv. This also simplifies its multithreading, eliminating the need for a complex, class-based pipeline.

BREAKING CHANGE: Real-world tests showed that it was quite challenging, if not impossible, to create two excel files that could be compared on a byte-by-byte level and be identical, even if their content was the same by eye. To obliviate this, and easing pipeline testing, the decision was made to move to the more reliable (and transparent) csv format.