test(multithreading): expand coverage of _get_session
Tests now also test that a new session is only returned if one does not exist. While this does not yet guarantee thread safety, it does ensure that each thread is at least continually using the same session.

It remains to be tested that each thread is using a different session, not just reusing the same session.