feat(request): do not write file if gene not found
This changes lut_check to retune the gene name if not found. Then, by checking if the gene starts with "ENSG" - a pattern that only occurs in Ensembl IDs, the decision can easily be made about whether os not to write a file.

BREAKING CHANGE: gtex_request will no longer always output a file. If the passed geneID does not begin with "ENSG", a warning will be logged and the request skipped. This designed to be used with snakemake checkpoints to allow unknown variable number of genes.